- Adding L0 sorting.  - Adding L0 carousel.  - Adding L1 sorting and  categorize contests.  - Adding practise coin chip icon.  - Adding create team assets.  - Adding user name and account balance to appdrawer.  - Adding innings name to my contest innings contest.  - Adding playing style icon to  save team, view team.  - Switch team after search contest was not working.  - Brand clip was visible for all contests insted of recommended.  - Saving initdata to shared preferences.

+++ lib/modal/l1.dart
@@ -296,7 +296,7 @@ class Contest {
   int prizeType;
   int entryFee;
   int minUsers;
-  int serviceFee;
+  double serviceFee;
   int teamsAllowed;
   int leagueId;
   int releaseTime;
@@ -359,7 +359,7 @@ class Contest {
       prizeType: json["prizeType"],
       entryFee: json["entryFee"],
       minUsers: json["minUsers"],
-      serviceFee: json["serviceFee"],
+      serviceFee: (json["serviceFee"]).toDouble(),
       teamsAllowed: json["teamsAllowed"],
       leagueId: json["leagueId"],
       releaseTime: json["releaseTime"],
